Ensure a sensors is updated after each read.
The energy sensors based on an output (ending with _wh) may not contain the correct result due to how the iotawatt is calculating them (such as a typical Import and Export energy rule).
It may be required to use the `integration` integration to calculate the energy the power device used over a given amount of time.
For the integration to rule, the sensor state must be updated frequently.
The Home Assistant value is to only update the value of a sensor if a change occurred which isn't what we want.

So we configure the HA DataUpdateCoordinator for forced updates.

Fixes #15