Automatically add an integration entity of power output sensor.
Querying the iotawatt for accumulated energy of a power output sensor will typically return the wrong value due to how the calculations are performed.

So rather than relying on the Iotawatt to perform those calculations we instead calculate the energy value locally by using the Riemann Sum Integration. The energy will be calculated after each read (performed every 30s)

The generated sensor will have a name in the format:
"IotaWatt Output Sensor Integral".

One downside over the energy sensor (ending with .wh) is that the energy will not accumulate following a reboot and the data will be lost.

This could be improved by querying the IotaWatt for the average power since it was last queried.

Fixes #18