Creates a new *.accumulated for each power entities. Those are populated by reading at regular intervals the energy value from the iotawatt. When declaring outputs relying on a mathematical expression (using other than + and - operator), the default energy sensors can't be used
While the `integration` integration could be used, the advantage of those sensors is that they won't lose data if the connection to the iotawatt is lost or if rebooting the HA instance.
In these cases, the sensor will fetch the data from when it last stop.

This change requires iotawattpy 0.1.0

Fixes gtdiehl#18