Give output sensors a unique name so that they show under the IoTaWatt integration.
Warning this is a breaking change in that it will add the new editable sensors on top of the existing ones.
To get around this and to not have to delete and rename anything do the following before updating.

Delete the integration. Restart HA.
Re-add the integration.

Now all your sensors will be there, no naming conflicts